Refactor SketchCanvas draft helpers
